Refactoring the search screen
