+++ IngrediScan/View/SearchScreen/test.swift
@@ -1,0 +1,7 @@
+//
+//  test.swift
+//  IngrediScan
+//
+//  Created by Jan Keller on 21.05.25.
+//
+
